requirements.txt
# Core Application Dependencies - ACTUAL INSTALLED VERSIONS
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.1
numpy==2.3.1
sqlalchemy==2.0.41
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Oracle Database (to be installed when needed)
cx_Oracle==8.3.0

# Machine Learning (to be installed next)
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
spacy==3.7.2

# Development Tools (to be installed)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
requests==2.31.0
asyncio-mqtt==0.11.1
